+++ lumin/nn/models/blocks/head.py
@@ -53,9 +53,9 @@ class AbsHead(AbsBlock):
 class AbsMatrixHead(AbsHead):
     def __init__(self, cont_feats:List[str], vecs:List[str], feats_per_vec:List[str], row_wise:bool=True,
                  lookup_init:Callable[[str,Optional[int],Optional[int]],Callable[[Tensor],None]]=lookup_normal_init,
-                 lookup_act:Callable[[str],Any]=lookup_act, freeze:bool=False, **kargs):
+                 lookup_act:Callable[[str],Any]=lookup_act, freeze:bool=False, bn_class:Callable[[int],nn.Module]=nn.BatchNorm1d, **kargs):
         super().__init__(cont_feats=cont_feats, cat_embedder=None, lookup_init=lookup_init, freeze=freeze)
-        self.vecs,self.fpv,self.row_wise,self.lookup_act = vecs,feats_per_vec,row_wise,lookup_act
+        self.vecs,self.fpv,self.row_wise,self.lookup_act,self.bn_class = vecs,feats_per_vec,row_wise,lookup_act,bn_class
         self.n_v,self.n_fpv = len(self.vecs),len(self.fpv)
         self._build_lookup()
             
@@ -383,6 +383,7 @@ class InteractionNet(AbsMatrixHead):
         lookup_init: function taking choice of activation function, number of inputs, and number of outputs an returning a function to initialise layer weights.
         lookup_act: function taking choice of activation function and returning an activation function layer
         freeze: whether to start with module parameters set to untrainable
+        bn_class: class to use for BatchNorm, default is `nn.BatchNorm1d`
     
     Examples::
         >>> inet = InteractionNet(cont_feats=matrix_feats, feats_per_vec=feats_per_vec,vecs=vecs,
@@ -404,9 +405,9 @@ class InteractionNet(AbsMatrixHead):
                  outfunc_depth:int, outfunc_width:int, outfunc_out_sz:int, agg_method:str,
                  do:float=0, bn:bool=False, act:str='relu',
                  lookup_init:Callable[[str,Optional[int],Optional[int]],Callable[[Tensor],None]]=lookup_normal_init,
-                 lookup_act:Callable[[str],Any]=lookup_act, freeze:bool=False, **kargs):
+                 lookup_act:Callable[[str],Any]=lookup_act, freeze:bool=False, bn_class:Callable[[int],nn.Module]=nn.BatchNorm1d, **kargs):
         super().__init__(cont_feats=cont_feats, vecs=vecs, feats_per_vec=feats_per_vec, row_wise=False,
-                         lookup_act=lookup_act, lookup_init=lookup_init, freeze=freeze)
+                         lookup_act=lookup_act, lookup_init=lookup_init, freeze=freeze, bn_class=bn_class)
         self.intfunc_depth,self.intfunc_width,self.intfunc_out_sz = intfunc_depth,intfunc_width,intfunc_out_sz
         self.outfunc_depth,self.outfunc_width,self.outfunc_out_sz = outfunc_depth,outfunc_width,outfunc_out_sz
         self.do,self.bn,self.act = do,bn,act
@@ -439,7 +440,7 @@ class InteractionNet(AbsMatrixHead):
         self.lookup_init(self.act, fan_in, fan_out)(layers[-1].weight)
         nn.init.zeros_(layers[-1].bias)
         if self.act != 'linear': layers.append(self.lookup_act(self.act))
-        if self.bn:  layers.append(nn.BatchNorm1d(fan_out))
+        if self.bn:  layers.append(LCBatchNorm1d(self.bn_class(fan_out)))
         if self.do: 
             if self.act == 'selu': layers.append(nn.AlphaDropout(self.do))
             else:                  layers.append(nn.Dropout(self.do))
@@ -622,6 +623,7 @@ class AbsConv1dHead(AbsMatrixHead):
         layer_kargs: dictionary of keyword arguments which are passed to `get_layers`
         lookup_init: function taking choice of activation function, number of inputs, and number of outputs an returning a function to initialise layer weights.
         freeze: whether to start with module parameters set to untrainable
+        bn_class: class to use for BatchNorm, default is `nn.BatchNorm1d`
     
     Examples::
         >>> class MyCNN(AbsConv1dHead):
@@ -667,8 +669,9 @@ class AbsConv1dHead(AbsMatrixHead):
     def __init__(self, cont_feats:List[str], vecs:List[str], feats_per_vec:List[str],
                  act:str='relu', bn:bool=False, layer_kargs:Optional[Dict[str,Any]]=None,
                  lookup_init:Callable[[str,Optional[int],Optional[int]],Callable[[Tensor],None]]=lookup_normal_init,
-                 lookup_act:Callable[[str],Any]=lookup_act, freeze:bool=False, **kargs):
-        super().__init__(cont_feats=cont_feats, vecs=vecs, feats_per_vec=feats_per_vec, row_wise=False, lookup_init=lookup_init, lookup_act=lookup_act, freeze=freeze)
+                 lookup_act:Callable[[str],Any]=lookup_act, freeze:bool=False, bn_class:Callable[[int],nn.Module]=nn.BatchNorm1d, **kargs):
+        super().__init__(cont_feats=cont_feats, vecs=vecs, feats_per_vec=feats_per_vec, row_wise=False, lookup_init=lookup_init, lookup_act=lookup_act,
+                         freeze=freeze, bn_class=bn_class)
         if layer_kargs is None: layer_kargs = {}
         self.layers:nn.Module = self.get_layers(in_c=self.n_fpv, act=act, bn=bn, **layer_kargs)
         self.out_sz = self.check_out_sz()
@@ -694,7 +697,7 @@ class AbsConv1dHead(AbsMatrixHead):
         if training: self.train()
         return x.size(-1)
             
-    def get_conv1d_block(self, in_c:int, out_c:int, kernel_sz:int, padding:Union[int,str]='auto', stride:int=1,act:str='relu', bn:bool=False) -> Conv1DBlock:
+    def get_conv1d_block(self, in_c:int, out_c:int, kernel_sz:int, padding:Union[int,str]='auto', stride:int=1, act:str='relu', bn:bool=False) -> Conv1DBlock:
         r'''
         Wrapper method to build a :class:`~lumin.nn.models.blocks.conv_blocks.ConvBlock` object.
 
@@ -714,7 +717,7 @@ class AbsConv1dHead(AbsMatrixHead):
         '''
         
         return Conv1DBlock(in_c=in_c, out_c=out_c, kernel_sz=kernel_sz, padding=padding, stride=stride, act=act, bn=bn,
-                           lookup_act=self.lookup_act, lookup_init=self.lookup_init)
+                           lookup_act=self.lookup_act, lookup_init=self.lookup_init, bn_class=self.bn_class)
     
     def get_conv1d_res_block(self, in_c:int, out_c:int, kernel_sz:int, padding:Union[int,str]='auto', stride:int=1,act:str='relu', bn:bool=False) -> Res1DBlock:
         r'''
@@ -736,7 +739,7 @@ class AbsConv1dHead(AbsMatrixHead):
         '''
 
         return Res1DBlock(in_c=in_c, out_c=out_c, kernel_sz=kernel_sz, padding=padding, stride=stride, act=act, bn=bn,
-                          lookup_act=self.lookup_act, lookup_init=self.lookup_init)
+                          lookup_act=self.lookup_act, lookup_init=self.lookup_init, bn_class=self.bn_class)
     
     def get_conv1d_resNeXt_block(self, in_c:int, inter_c:int, cardinality:int, out_c:int, kernel_sz:int, padding:Union[int,str]='auto', stride:int=1,
                                  act:str='relu', bn:bool=False) -> ResNeXt1DBlock:
@@ -761,7 +764,7 @@ class AbsConv1dHead(AbsMatrixHead):
         '''
 
         return ResNeXt1DBlock(in_c=in_c, inter_c=inter_c, cardinality=cardinality, out_c=out_c, kernel_sz=kernel_sz, padding=padding, stride=stride, act=act,
-                              bn=bn, lookup_act=self.lookup_act, lookup_init=self.lookup_init)
+                              bn=bn, lookup_act=self.lookup_act, lookup_init=self.lookup_init, bn_class=self.bn_class)
     
     @abstractmethod
     def get_layers(self, in_c:int, act:str='relu', bn:bool=False, **kargs) -> nn.Module:
@@ -838,6 +841,7 @@ class LorentzBoostNet(AbsMatrixHead):
             Purely for inheritance, unused by class as is.
         lookup_act: function taking choice of activation function and returning an activation function layer. Purely for inheritance, unused by class as is.
         freeze: whether to start with module parameters set to untrainable.
+        bn_class: class to use for BatchNorm, default is `nn.BatchNorm1d`
     
     Examples::
         >>> lbn = LorentzBoostNet(cont_feats=matrix_feats, feats_per_vec=feats_per_vec,vecs=vecs, n_particles=6)
@@ -852,15 +856,16 @@ class LorentzBoostNet(AbsMatrixHead):
     def __init__(self, cont_feats:List[str], vecs:List[str], feats_per_vec:List[str],
                  n_particles:int, feat_extractor:Optional[Callable[[Tensor],Tensor]]=None, bn:bool=True,
                  lookup_init:Callable[[str,Optional[int],Optional[int]],Callable[[Tensor],None]]=lookup_normal_init,
-                 lookup_act:Callable[[str],Any]=lookup_act, freeze:bool=False, **kargs):
-        super().__init__(cont_feats=cont_feats,vecs=vecs,feats_per_vec=feats_per_vec,row_wise=True,lookup_init=lookup_init,lookup_act=lookup_act,freeze=freeze)
+                 lookup_act:Callable[[str],Any]=lookup_act, freeze:bool=False, bn_class:Callable[[int],nn.Module]=nn.BatchNorm1d, **kargs):
+        super().__init__(cont_feats=cont_feats,vecs=vecs,feats_per_vec=feats_per_vec,row_wise=True,lookup_init=lookup_init,lookup_act=lookup_act,freeze=freeze,
+                         bn_class=bn_class)
         self.n_particles = n_particles
         self.comb = torch.combinations(torch.arange(0,self.n_particles))
         if feat_extractor is not None: self.feat_extractor = feat_extractor
         self.part_wgts,self.rf_wgts = self._get_wgts(),self._get_wgts()
         self.out_sz = None
         self.out_sz = self.check_out_sz()
-        self.bn = nn.BatchNorm1d(self.out_sz) if bn else None
+        self.bn = self.bn_class(self.out_sz) if bn else None
         self._map_outputs()
         if self.freeze: self.freeze_layers()
     
@@ -917,7 +922,7 @@ class LorentzBoostNet(AbsMatrixHead):
         Passes input through the LB network and aggregates down to a flat tensor via the feature extractor, optionally passing through a batchnorm layer.
 
         Arguments:
-            x: If a tuple, the second element is assumed to the be the matrix data. If a flat tensor, will conver the data to a matrix
+            x: If a tuple, the second element is assumed to the be the matrix data. If a flat tensor, will convert the data to a matrix
         
         Returns:
             Resulting tensor
@@ -991,6 +996,7 @@ class AutoExtractLorentzBoostNet(LorentzBoostNet):
         lookup_init: function taking choice of activation function, number of inputs, and number of outputs an returning a function to initialise layer weights.
         lookup_act: function taking choice of activation function and returning an activation function layer.
         freeze: whether to start with module parameters set to untrainable.
+        bn_class: class to use for BatchNorm, default is `nn.BatchNorm1d`
     
     Examples::
         >>> aelbn = AutoExtractLorentzBoostNet(cont_feats=matrix_feats, feats_per_vec=feats_per_vec,vecs=vecs, n_particles=6,
@@ -1000,7 +1006,7 @@ class AutoExtractLorentzBoostNet(LorentzBoostNet):
     def __init__(self, cont_feats:List[str], vecs:List[str], feats_per_vec:List[str],
                  n_particles:int, depth:int, width:int, n_singles:int=0, n_pairs:int=0, act:str='swish', do:float=0, bn:bool=False, 
                  lookup_init:Callable[[str,Optional[int],Optional[int]],Callable[[Tensor],None]]=lookup_normal_init,
-                 lookup_act:Callable[[str],Any]=lookup_act, freeze:bool=False, **kargs):
+                 lookup_act:Callable[[str],Any]=lookup_act, freeze:bool=False, bn_class:Callable[[int],nn.Module]=nn.BatchNorm1d, **kargs):
         self.n_singles,self.n_pairs = n_singles,n_pairs
         
         # Mock NNs to allow out_sz computation
@@ -1010,13 +1016,13 @@ class AutoExtractLorentzBoostNet(LorentzBoostNet):
         self.pre_bn    = lambda x: x
         
         super().__init__(cont_feats=cont_feats, vecs=vecs, feats_per_vec=feats_per_vec, n_particles=n_particles,
-                         bn=False, lookup_init=lookup_init, lookup_act=lookup_act, freeze=freeze)
+                         bn=False, lookup_init=lookup_init, lookup_act=lookup_act, freeze=freeze, bn_class=bn_class)
         
         if n_singles > 0: self.single_nn = self._get_nn(n_in=4, depth=depth, width=width,n_out=n_singles, act=act, do=do, bn=bn,
                                                         lookup_act=lookup_act, lookup_init=lookup_init)
         if n_pairs   > 0: self.pair_nn   = self._get_nn(n_in=8, depth=depth, width=width, n_out=n_pairs, act=act, do=do, bn=bn,
                                                         lookup_act=lookup_act, lookup_init=lookup_init)
-        self.pre_bn = nn.BatchNorm1d(4*self.n_particles)
+        self.pre_bn = self.bn_class(4*self.n_particles)
     
     def _get_nn(self, n_in:int, depth:int, width:int, n_out:int, act:str, do:bool, bn:bool,
                 lookup_init:Callable[[str,Optional[int],Optional[int]],Callable[[Tensor],None]],
@@ -1032,7 +1038,7 @@ class AutoExtractLorentzBoostNet(LorentzBoostNet):
         lookup_init(act, n_in, n_out)(layers[-1].weight)
         nn.init.zeros_(layers[-1].bias)
         if act != 'linear': layers.append(lookup_act(act))
-        if bn:  layers.append(nn.BatchNorm1d(n_out))
+        if bn:  layers.append(LCBatchNorm1d(nn.BatchNorm1d(n_out)))
         if do: 
             if act == 'selu': layers.append(nn.AlphaDropout(do))
             else:             layers.append(nn.Dropout(do))
@@ -1098,7 +1104,7 @@ class GravNet(AbsMatrixHead):
         lookup_init: function taking choice of activation function, number of inputs, and number of outputs an returning a function to initialise layer weights.
         lookup_act: function taking choice of activation function and returning an activation function layer
         freeze: whether to start with module parameters set to untrainable
-        bn_class: class to use for BatchNorm, default is :class:`~lumin.nn.models.layers.batchnorms.LCBatchNorm1d`
+        bn_class: class to use for BatchNorm, default is `nn.BatchNorm1d`
     '''
 
     def __init__(self, cont_feats:List[str], vecs:List[str], feats_per_vec:List[str],
@@ -1109,14 +1115,14 @@ class GravNet(AbsMatrixHead):
                  agg_methods:Union[List[str],str]=['mean','max'], concat_outs:bool=True, flatten:bool=False,
                  do:float=0, bn:bool=False, act:str='relu',
                  lookup_init:Callable[[str,Optional[int],Optional[int]],Callable[[Tensor],None]]=lookup_normal_init,
-                 lookup_act:Callable[[str],Any]=lookup_act, freeze:bool=False, bn_class:Callable[[int],nn.Module]=LCBatchNorm1d, **kargs):
+                 lookup_act:Callable[[str],Any]=lookup_act, freeze:bool=False, bn_class:Callable[[int],nn.Module]=nn.BatchNorm1d, **kargs):
         super().__init__(cont_feats=cont_feats, vecs=vecs, feats_per_vec=feats_per_vec, row_wise=True,
-                         lookup_act=lookup_act, lookup_init=lookup_init, freeze=freeze)
-        store_attr(but=[cont_feats, vecs, feats_per_vec, lookup_act, lookup_init, freeze, agg_methods])
+                         lookup_act=lookup_act, lookup_init=lookup_init, freeze=freeze, bn_class=bn_class)
+        store_attr(but=[cont_feats, vecs, feats_per_vec, lookup_act, lookup_init, freeze, agg_methods, bn_class])
         self._check_agg_method(agg_methods)
         if not is_listy(self.n_out): self.n_out = [self.n_out]
         
-        if self.use_in_bn: self.in_bn = self.bn_class(self.n_fpv)
+        if self.use_in_bn: self.in_bn = LCBatchNorm1d(self.bn_class(self.n_fpv))
         self.grav_layers = self._get_grav_layers()
         self.f_final = self._get_final_layers()
         self._map_outputs()
@@ -1146,7 +1152,7 @@ class GravNet(AbsMatrixHead):
         self.lookup_init(self.act, fan_in, fan_out)(layers[-1].weight)
         nn.init.zeros_(layers[-1].bias)
         if self.act != 'linear': layers.append(self.lookup_act(self.act))
-        if self.bn:              layers.append(self.bn_class(fan_out))
+        if self.bn:              layers.append(LCBatchNorm1d(self.bn_class(fan_out)))
         if self.do: 
             if self.act == 'selu': layers.append(nn.AlphaDropout(self.do))
             else:                  layers.append(nn.Dropout(self.do))